tests: link against libunwind, now that non-fake libunwind is default
In 49afa666 we reverted to using stock libunwind instead of our
hacked-up (but super-fast) frame-pointer-only version. The right fix is
to use a fast-but-accurate unwinder (as I have been working on with
Théophile and Francesco in the dwarf-assembly project). In the meantime,
fix the build of test cases by linking against libunwind. Being a quick
fix, this is an arch-specific HACK for now.