Tweak compiler wrapper to build with relocs but then strip them out
The idea here is that rather than retain spent relocs at run time, we
build with them included but, once we've run metavector, we strip them
out. The actual strip command is commented out for now... largely for
debugging, but since it breaks the usual property that we can rebuild
metadata from the binary on disk (using Makefile.meta) it may be better
to rethink this.