tools/uniqtypes.cpp: mostly revert d027b4b1 (section flags hack)
This was an alternative scheme for __attribute__-based section
flags/groups, but it doesn't seem to work so I've reverted it. Sections
were not being re-used after declaration, but were duplicated and the
duplicates were not included in the group, causing link-time multiple
definition bugs in meta-DSOs and perhaps bugs of non-aliasing in other
cases. Also, the motivation for this change (that recent GCC demands
that aliases' sections match their referents') was already handled by
ensuring aliases' attribute strings were a verbatim match (again, works
for me) and the portability aspect was not yet tested (it would be nice
if our meta-objects could be compiled with Clang, but we had no test for
that).