Rule out shared libraries with holes in, by loading the ld.so ourselves
Shared libraries mapped with (non-empty) holes in their virtual address
space break our allocation hierarchy invariant (see GitHub issue #47).
Since on glibc-based GNU/Linux the only library that can be mapped this
way is the ld.so (mapped by the kernel, possibly holeyly), create a new
ld.so that chain-loads the ld.so but mapped *without* holes, and make
allocscc build binaries that use it as their interpreter.