Revert "liballocs_cil_inlines.h: fix version-skewed macros after previous commit cleanup"
This reverts commit 1c62e914 and reinstates some macros from the high water mark.
This retains compatibility with libcrunch's inlines, which use some of the same
macros. It is good to have a uniform set of macros for cache-related tests, and
these versions take enough arguments that they can be replaced with inline
functions if we need to do that (as when investigating the Heisenbug).