Configuration fixes
Set DONALD var analogously with LIBRUNT. Probe for needed boost
libraries. Export LIBRUNT so test cases can see it, and use it in
allocscompilerwrapper.py when compiling allocator stubs. Remove dwarf.h
from our configure-time requirements, so it may be provided as a DWARF_H
preprocessor macro (a bit of a hack).