allocsld: add a new track-covering trick for debugging
If gdb starts the program itself, it will try to breakpoint
_dl_debug_state() (or whatever) on the ld.so that the program binary
requests, identified by looking at the binary on disk rather than the
PT_INTERP string in memory. So we need to provide this function
ourselves, even though we make it look like the real ld.so is in charge.
We very nastily clobber the real ld.so's version of this function so
that it jumps to our own. Breakpointing either one should be sufficient
for debugging to work.