Rejig the generic malloc allocator so that promoted chunks are still indexed.
This allows the allocator to walk all its chunks in order, and will help
simplify the way we walk the allocation tree depth-first.

One wart right now: get_info will not succeed if given pointers far within the
object, because it still bounds its search using the size of the biggest
*non-promoted* chunk.

Also this breaks Guillaume's lifetime policies stuff. I couldn't cope
with the complexity of two kinds of insert. Instead we'll revisit this
when we free up some spare bits in the insert, on moving from a memtable
to a bitmap.