Eliminate the generic_malloc allocator, moving its indexing routines to a header
Instead of one catch-all allocator identity for malloc-likes, we want to
split them up, to allow their behaviours (e.g. alignment) and indexing
code to differ. Pull the bitmap-based indexing routines into a header
and define a __default_lib_malloc_allocator to represent the preloadable
'malloc'. This will usually be the global malloc; define
__global_malloc_allocator weakly aliasing it. This breaks the
malloc-in-exe test case, for now, since we need to generate a separate
'struct allocator' instance at the same time as when we link in its
special (wrapdl) wrappers.