default-lib-malloc.c: macroise the definitions
The indexing glue is ready to be move to stubgen.h, but it is staying
put for now. This also cleaned up some cruft around lifetime policies
and 'struct insert'. Eventually we want to get rid of 'struct
extended_insert' and just scavenge enough bits from the insert proper.
Since policies are not allocators, but meta-rules for how allocators'
requests ("free this now") should compose, a few bits are enough.