Add test case of calling malloc *from* a DSO, also cleaning up linkage
Default lib malloc's event hooks are public not hidden, but always with
the symname prefix __liballocs_malloc_. We could use a hidden alias
internally, but the global symname must be available because the events
may occur in a DSO that's not liballocs (e.g. in the executable, if it
preloads malloc, or in a hypothetical libc that is already instrumented).