Eliminate nonshared malloc stubs and consoliate malloc_usable_size
Here we eliminate the horrible mess of malloc_usable_size variants, by
parameterising the generic malloc indexing routines by their size
function. Since these are intended to be inlined, the call should not
stay an indirect call in the generated code.

The nonshared library no longer has a reason to exist. The same goes for
the liballocs.so linker script, although that remains for now.

This leaves broken tests malloc-in-exe (still) and malloc-in-dso
(newly). I will overhaul stub generation shortly to fix these.

Also fix some documentation and comments.