Plumbing for ld.so crash workarounds
We are resigned to working around ld.so bugs such as glibc's bug 31168,
by introspecting on its corrupt data structures. Also we want ld.so
allocations to be introspectable generally. So here we add some
foundations for this. This includes a basic binary patcher that can
catch __minimal_malloc and the like, as used by glibc's ld.so. It's not
currently enabled, and needs to be refined somewhat.