Add ability to introspect on ld.so malloc allocations
This required addindg a new 'ld-so-malloc' allocator and generalising
the stubgen macros a bit. We have a Detours-style binary instrumentation
system in allocsld for doing this.

There is also some stuff in workarounds.c and the static-symbol
allocator about reflective access to data structures (like the ld.so's
link map structures). This will likely go away because we're no longer
pursuing the workaround of these up at load time. However, the code is
still there for now and may be repurposable.

The new "linear malloc" index probably needs generalising so that we can
remember events happening during init (like we also do with early_libs)
and "play them back" when we're ready to process them. See GitHub issue
\#97.