Add initial version of 'frametypes2' tool
This enhanced version of frametypes can generate more complete stack
frame information than the current one. In particular it can describe
what is held in registers and can describe the contents of save slots
mentioned in the CFI (.eh_frame or .debug_frame). The idea is to extend
it even further: to computed values, and possibly to plug gaps during
function prologues. Also the output representation is to be much more
compact, since it will not involve uniqtypes for the whole frame. The
current version doesn't output anything yet....