Revert ".circleci: experiment with 22.04 + gcc-10"
This reverts commit bd5a6fb8757a529d4ee58bec30b0830227607f78. The
CircleCI build fails with a lack of 'g++' which is easily fixed by
symlink. The bigger problem is that although we could build liballocs
just fine in this environment, the glibc is built with gcc 11 and so
uses DWARF 5, so the metadata build will fail. That could be worked
around by rebuilding glibc as part of the liballocs build, but that's
nasty and time-consuming. For now we just stick at 20.04.