Split most of liballocs.c out into various smaller compilation units
The file liballocs.c is really a hangover from when the library was much
simpler. Splitting it up in this way allows allocsld to borrow some of
its functionality (here in allocinstr.c, for loading meta-DSOs), albeit
not without hacks for now (around TLS, and avoiding pulling in
problematic parts of librunt).

Of course it should also make the code more comprehensible. Soon
liballocs.c will be renamed to something like init.c, since only
initialization code is left there.