tools/allocs-wrapper: don't embed useless quotes in -cflags/-ldflags mode
The idea of allocs-cflags / allocs-ldflags is that we use them like
'pkg-config', e.g. CFLAGS += `/path/to/allocs-cflags`. However, this is
doomed not to work! Unless IFS is changed, *any* embedded space in the
allocs-cflags output, whether quoted/escaped or not, will trigger
word-splitting. The right way, in a Makefile context, is probably to do
something like (ugly!):

CFLAGS += $(shell IFS='\n' /path/to/allocs-cflags)

... and have the allocs-cflags print newline-separated words, i.e. it's
OK to drop support for embedded newlines, whereas for embedded spaces
this is not OK. TODO: find any pkg-config recipes for the same problem.