Make interp-pad.o to avoid depending on gold or BFD linker behaviour
Since gold and BFD linkers diverge on what it means to link in your own
.interp section (gold clobbers -dynamic-linker, BFD appends to it), use
string merging to erase the difference. Also update donald to a version
that can grok a multi-string .interp section and make
allocscompilerwrapper.py set the dynamic linker using exactly the same
string as appears first in inter-pad.o.