Refactor alloca handling so that we no longer wrap alloca
Wrapping alloca() in an inlined helper confuses Clang. Strictly speaking
the allocated alloca chunk's lifetime ends at the end of the inline. GCC
promotes its lifetime to that of the enclosing physical frame, which is
sensible, but Clang seems not to. Here we refactor the instrumentation
so that it takes the form of pre- and post-calls (size helper and
notifier, respectively) rather than a wrapper.