Preserve ordering for `-mllvm <arg>`
We had a report from Marc at Galois that the ordering of arguments got
tripped when e.g. `-mllvm --blarg` is passed through the compiler
wrapper.

This change ensures the ordering of flag and argument is preserved.

However, there's a general case of this: we don't know which `-m`,
`-f` etc args work like flags and which ones accept values so it'll
continue to pop up until we ensure we always preserve the ordering.
I _think_ the ordering is usually preserved and the compiler wrapper is
getting tripped up in the case of `-mllvm` because it takes another flag
as a value (i.e. `--blarg` instead of `blarg`). I'm unsure. If that _is_
what's tripping it up, it should be a rare case, so this fix should
suffice for now.